+++ tests/common/Tag/H4Test.php
@@ -14,7 +14,7 @@ final class H4Test extends TestCase
         $this->assertSame(
             '<h4 class="red">Hello</h4>',
             (string) H4::tag()
-                ->class('red')
+                ->replaceClass('red')
                 ->content('Hello')
         );
     }