+++ tests/common/Tag/H6Test.php
@@ -14,7 +14,7 @@ final class H6Test extends TestCase
         $this->assertSame(
             '<h6 class="red">Hello</h6>',
             (string) H6::tag()
-                ->class('red')
+                ->replaceClass('red')
                 ->content('Hello')
         );
     }